Build an actual NROM-128 rom
Everything was configured as NROM-128, with one PRG and one CHR, but
for some reason I was duplicating the PRG as NROM-256.
(Ok, "some reason" was ChatGPT dreaming.)

This removes the duplicate PRG in $8000-$bfff, and the NES (or emulator)
would naturally duplicate $C000-$FFFF into those spaces.

The "define=yes" labels in the .cfg were also removed.  We don't
reference nor need those labels.